requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
httpie==2.3.0
idna==2.10
Pygments==2.7.4
PyJWT==2.0.1
PySocks==1.7.1
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.3
